builtin.system_init -> core.system_init
Move system initialization module out of mathics.builtin.
(Similar to PR #639)

This also removes the need for Python to "partially initialize"
mathics.builtin since it doesn't import one of its children

Merge mess
