+++ mathics/session.py
@@ -12,11 +12,11 @@ In particular we provide:
 import os.path as osp
 from typing import Optional
 
-from mathics.builtin.system_init import initialize_system
 from mathics.core.definitions import autoload_files
 from mathics.core.parser import parse, MathicsSingleLineFeeder
 from mathics.core.definitions import Definitions
 from mathics.core.evaluation import Evaluation
+from mathics.core.system_init import initialize_system
 import mathics.settings
